soil-backend/requirements.txt
fastapi
uvicorn[standard]
ultralytics
pillow
python-multipart
pandas
numpy
scikit-learn
openpyxl
python-dotenv
supabase
python-jose
requests
